# docs/api/openapi.yml
openapi: 3.0.3
info:
  title: Algorithmic Trading System API
  description: |
    Enterprise-grade algorithmic trading platform API providing comprehensive trading, portfolio management, 
    risk management, and market data services. Built with microservices architecture and event-driven design.
    
    ## Key Features
    - **Multi-Asset Trading**: Stocks, ETFs, Options, Futures, Forex, Crypto
    - **Real-time Market Data**: Live feeds with fallback mechanisms
    - **Advanced Risk Management**: Real-time monitoring, VaR, stress testing
    - **AI-Powered Analytics**: Strategy development and optimization
    - **High-Performance Execution**: Microsecond-level latency
    
    ## Authentication
    All API endpoints require authentication via JWT tokens or API keys.
    
    ## Rate Limiting
    - Standard endpoints: 1000 requests/minute
    - Market data endpoints: 10000 requests/minute
    - Trading endpoints: 500 requests/minute
    
    ## Error Handling
    The API uses standard HTTP status codes and returns detailed error information in JSON format.
  version: 1.0.0
  contact:
    name: Trading System API Support
    email: api-support@trading-system.com
    url: https://docs.trading-system.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.trading-system.com/v1
    description: Production server
  - url: https://staging-api.trading-system.com/v1
    description: Staging server
  - url: http://localhost:8000/v1
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: User login
      description: Authenticate user and return JWT token
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /market-data/quotes/{symbol}:
    get:
      tags: [Market Data]
      summary: Get real-time quote
      description: Retrieve real-time quote for a specific symbol
      operationId: getQuote
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
            example: "AAPL"
          description: Trading symbol
      responses:
        '200':
          description: Quote retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quote'
        '404':
          description: Symbol not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /orders:
    get:
      tags: [Trading]
      summary: Get orders
      description: Retrieve user's orders with optional filtering
      operationId: getOrders
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, filled, cancelled, rejected]
        - name: symbol
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Orders retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'

    post:
      tags: [Trading]
      summary: Place order
      description: Place a new trading order
      operationId: placeOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequest'
      responses:
        '201':
          description: Order placed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid order parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /orders/{order_id}:
    get:
      tags: [Trading]
      summary: Get order details
      description: Retrieve details of a specific order
      operationId: getOrder
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Trading]
      summary: Cancel order
      description: Cancel a pending order
      operationId: cancelOrder
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Order cannot be cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Portfolio
...
  /portfolio/performance:
    get:
      tags: [Portfolio]
      summary: Get performance metrics
      description: Retrieve portfolio performance metrics
      operationId: getPerformance
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [1d, 1w, 1m, 3m, 6m, 1y, ytd, all]
            default: 1m
      responses:
        '200':
          description: Performance metrics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PerformanceMetrics'

  # Risk Management
...
  /strategies:
    get:
      tags: [Strategies]
      summary: Get trading strategies
      description: Retrieve list of available trading strategies
      operationId: getStrategies
      responses:
        '200':
          description: Strategies retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Strategy'

    post:
      tags: [Strategies]
      summary: Create strategy
      description: Create a new trading strategy
      operationId: createStrategy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StrategyRequest'
      responses:
        '201':
          description: Strategy created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Strategy'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
...
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          default: bearer
        expires_in:
          type: integer
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, trader, analyst, viewer]
        created_at:
          type: string
          format: date-time

    # Market Data Schemas
    Quote:
      type: object
      properties:
        symbol:
          type: string
        bid:
          type: number
        ask:
          type: number
        last:
          type: number
        volume:
          type: integer
        timestamp:
          type: string
          format: date-time
        market_status:
          type: string
          enum: [open, closed, pre_market, after_hours]
...
    OrderRequest:
      type: object
      required: [symbol, side, quantity, order_type]
      properties:
        symbol:
          type: string
        side:
          type: string
          enum: [buy, sell]
        quantity:
          type: number
          minimum: 0
        order_type:
          type: string
          enum: [market, limit, stop, stop_limit]
        price:
          type: number
          minimum: 0
        stop_price:
          type: number
          minimum: 0
        time_in_force:
          type: string
          enum: [day, gtc, ioc, fok]
          default: day
        client_order_id:
          type: string
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        client_order_id:
          type: string
        symbol:
          type: string
        side:
          type: string
          enum: [buy, sell]
        quantity:
          type: number
        filled_quantity:
          type: number
        remaining_quantity:
          type: number
        order_type:
          type: string
          enum: [market, limit, stop, stop_limit]
        price:
          type: number
        stop_price:
          type: number
        status:
          type: string
          enum: [pending, filled, partially_filled, cancelled, rejected]
        time_in_force:
          type: string
          enum: [day, gtc, ioc, fok]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        fills:
          type: array
          items:
            $ref: '#/components/schemas/Fill'
    Fill:
      type: object
      properties:
        id:
          type: string
          format: uuid
        quantity:
          type: number
        price:
          type: number
        timestamp:
          type: string
          format: date-time
        commission:
          type: number

    # Portfolio Schemas
...
    PerformanceMetrics:
      type: object
      properties:
        total_return:
          type: number
        annualized_return:
          type: number
        volatility:
          type: number
        sharpe_ratio:
          type: number
        max_drawdown:
          type: number
        win_rate:
          type: number
        profit_factor:
          type: number

    # Risk Management Schemas
...
    Strategy:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [active, inactive, backtesting]
        parameters:
          type: object
        performance:
          $ref: '#/components/schemas/PerformanceMetrics'
        created_at:
          type: string
          format: date-time
    StrategyRequest:
      type: object
      required: [name, code]
      properties:
        name:
          type: string
        description:
          type: string
        code:
          type: string
        parameters:
          type: object
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: integer
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          format: uuid
...